requirements.txt
flask
python-dotenv
langchain-huggingface
langchain-core
pytesseract
Pillow